+++ CHANGELOG.md
@@ -6,17 +6,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [to be added]
-* Automatic distribution logo detection in place of/alongside left-side ascii art
 * Package count support
 * NO\_COLOR support
 * Refactor README
+
+## [1.4.0]
+* Added pfetch-like distro logo detection
+* Changed variable denoters from `|` to `$`
+* Removed ID\_LIKE support, instead default Linux logo/color is used for fallback
 
 ## [1.3.0]
 * Added command line options, including --help, --version and others
 * Added MEMUSED and SWAPUSED support
 
 ###  [1.2.1] - 2023-04-29
-* Added ID_LIKE support
+* Added ID\_LIKE support
 
 ##  [1.2.0] - 2023-04-21
 * Added fallback config